add file check extention and empty
