+++ file_parser/get_next_line_utils.c
